+++ src/main/java/controller/LottoController.java
@@ -2,13 +2,13 @@ package controller;
 
 import java.util.List;
 import java.util.function.Supplier;
-import model.BonusNumber;
-import model.Lotto;
-import util.LottoGenerator;
-import model.Purchase;
-import model.WinningNumber;
-import model.WinningResult;
-import model.WinningStatus;
+import model.draw.BonusNumber;
+import model.purchase.Lotto;
+import model.lottoNumber.LottoGenerator;
+import model.purchase.Purchase;
+import model.draw.WinningNumber;
+import model.draw.WinningResult;
+import model.draw.WinningStatus;
 import util.Parser;
 import view.InputView;
 import view.OutputView;
