+++ src/main/java/model/draw/WinningStatus.java
@@ -1,4 +1,4 @@
-package model;
+package model.draw;
 
 import java.util.Arrays;
 import java.util.Comparator;
