+++ src/main/java/model/lottoNumber/NumberPickStrategy.java
@@ -1,4 +1,4 @@
-package util;
+package model.lottoNumber;
 
 public interface NumberPickStrategy {
     int pickInRange(int min, int max);
