+++ src/main/java/model/lottoNumber/RandomNumberPickStrategy.java
@@ -1,4 +1,4 @@
-package util;
+package model.lottoNumber;
 
 import java.util.Random;
