+++ src/test/java/model/draw/BonusNumberTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.draw;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
