+++ src/test/java/model/draw/WinningResultTest.java
@@ -1,7 +1,8 @@
-package model;
+package model.draw;
 
 import static org.assertj.core.api.Assertions.*;
 
+import model.purchase.Purchase;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
