+++ src/test/java/model/lottoNumber/NumberGeneratorTest.java
@@ -1,4 +1,4 @@
-package util;
+package model.lottoNumber;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
